requirements.txt
opencv-python>=4.8.0
deepface>=0.0.79
numpy>=1.21.0
Pillow>=9.0.0 